Stabilize workspace-scoped trace archival coordination
Keep log_spans() and duplicate start_trace() retries on a single workspace
snapshot while simplifying trace_info mutations to update directly by trace_id
after workspace-scoped reads.

Fold trace_version into the unreleased workspace archival migration, use an
Integer counter, and distinguish deleted traces from archived/non-DB-backed
write conflicts in the failure path.

+++ mlflow/store/tracking/dbmodels/models.py
@@ -755,7 +755,7 @@ class SqlTraceInfo(Base):
     """
     Response preview: `String` (limit 1000 characters). Could be *null*. Newly added in V3 format.
     """
-    trace_version = Column(BigInteger, nullable=False, server_default="0")
+    trace_version = Column(Integer, nullable=False, server_default="0")
     """
     DB-backed trace payload generation used for concurrency coordination.
     Defaults to 0.
